index_elastic_float_price

+++ Modules/Prices/database/migrations/2024_03_26_191603_create_price_parse_table.php
@@ -6,16 +6,13 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('price_parse', function (Blueprint $table) {
             $table->id();
             $table->integer('price_uploaded_id')->unsigned()->index();
             $table->string('model');
-            $table->integer('price')->default(0);
+            $table->float('price')->default(0);
             $table->integer('quantity')->default(0);
             $table->longText('additional')->nullable();
             $table->integer('currency_id')->default(0)->unsigned()->index();
@@ -23,9 +20,6 @@ return new class extends Migration
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('price_parse');
