+++ Modules/Prices/database/migrations/2024_04_10_193046_create_inventories_table.php
@@ -6,16 +6,13 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('inventories', function (Blueprint $table) {
             $table->id();
             $table->integer('price_model_id');
             $table->integer('product_id');
-            $table->integer('price')->default(0);
+            $table->float('price')->default(0);
             $table->integer('currency_id')->default(0);;
             $table->integer('qty')->default(0);
             $table->integer('import_status')->default(0);
@@ -23,9 +20,6 @@ return new class extends Migration
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('inventories');
