+++ Modules/TradeZone/database/migrations/2024_04_10_200515_create_rules_table.php
@@ -6,25 +6,19 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('rules', function (Blueprint $table) {
             $table->id();
             $table->integer('is_active')->default(0);
-            $table->integer('price_min')->default(0);
-            $table->integer('price_max')->default(0);
+            $table->float('price_min')->default(0);
+            $table->float('price_max')->default(0);
             $table->text('trade_margin');
             $table->integer('sort')->default(0);
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('rules');
